added new Default command to add external assemblies TODO: Commands can only be invoked once, don't know reason yet
